src/requirements.txt
python-crontab==2.3.6
dash==0.35.1
dash-html-components==0.13.4
dash-core-components==0.42.1
dash-table==3.1.11